% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{cmgd_studies}
\alias{cmgd_studies}
\title{Studies in a release}
\usage{
cmgd_studies(rel)
}
\arguments{
\item{rel}{A release from \code{\link[=cmgd_release]{cmgd_release()}}.}
}
\value{
A data.frame with columns \code{study_name} and \code{n_samples}.
}
\description{
From the release's \code{studies/index.json}.
}
\examples{
\dontrun{
cmgd_studies(cmgd_release("cmgd_nextflow-2.2.1"))
}
}
